Update build process

+++ scripts/import_recipes.py
@@ -3,11 +3,16 @@ from pathlib import Path
 from string import Template
 from urllib.parse import urljoin
 
+import json
 import re
 import requests
 import shutil
 import unicodedata
 import yaml
+
+ROOT = Path(__file__).resolve().parent.parent
+
+ASSET_INDEX_PATH = ROOT / "astro/src/content/_asset-index.json"
 
 
 def remove_non_ascii_normalized(string: str) -> str:
@@ -21,6 +26,29 @@ def slugify(s):
 	s = re.sub(r'[\s_-]+', '-', s)
 	s = re.sub(r'^-+|-+$', '', s)
 	return s
+
+
+def add_asset(asset_index, slug, title, image_urls):
+	images = []
+	for img_url in image_urls:
+		filename = normalize_filename(img_url)
+		images.append({
+			"id": filename,
+			"source": img_url,
+			"filename": filename
+		})
+	asset_index["recipes"].append({
+		"slug": slug,
+		"title": title,
+		"images": images
+	})
+
+
+def normalize_filename(url: str):
+    return url.split("?")[0].split("/")[-1].strip()
+
+def write_asset_index(asset_index):
+    ASSET_INDEX_PATH.write_text(json.dumps(asset_index, indent=2))
 
 
 class Logger:
@@ -61,19 +89,29 @@ ${content}
 API = 'https://burnbookapi.fauno.nl/'
 IMAGE_OUTPUT_DIR = Path(__file__).absolute().parent / "output" / "images"
 
-def get_recipes():
-	url = f'{API}/recipes/'
+def get_recipes(logger):
+	url = urljoin(API, '/recipes/')
 	response = requests.get(url)
 	if response.status_code == 200:
 		return response.json()
 	else:
-		print('Request failed with status code', response.status_code)
+		logger.error(f'Failed to fetch recipes from API. Using {url}')
+		logger.info(f'Request failed with status code {response.status_code}')
+		logger.info(f'Response {response.text}')
 		exit(2)
 
 def download_image(image_url, output_folder, logger):
+	image_filename = image_url.split('/')[-1]
+	image_dir = Path(output_folder)
+	image_path = output_folder / image_filename
+	print(f"Downloading image from {image_url} to {image_path}")
+	print(f"Image exists {image_path.exists()}")
+	if image_path.exists():
+		logger.success(f"Skipping existing image: {image_path}")
+		return
+	
 	response = requests.get(image_url, stream=True)
 	if response.status_code == 200:
-		image_path = output_folder.joinpath(image_url.split('/')[-1])
 		with open(image_path, 'wb') as f:
 			shutil.copyfileobj(response.raw, f)
 		logger.success(f'Successfully downloaded image: {image_path}')
@@ -130,12 +168,25 @@ def run():
 	image_folder.mkdir(exist_ok=True)
 	logger.success('Successfully created output directories')
 
-	recipes = get_recipes()
+	asset_index = {
+		"recipes": []
+	}
+
+	recipes = get_recipes(logger)
 	template = Template(tpl)
 
 	for recipe in recipes:
 		# Remove instructions and add it as the body
 		content = recipe.get('instructions')
+		title = recipe.get('title', '')
+		image_urls = []
+		image_path = recipe.get('image', '')
+		if image_path:
+			image_url = urljoin(
+				API,
+				recipe.get('image', '')
+			)
+			image_urls.append(image_url)
 		del recipe['instructions']
 
 		frontmatter = (yaml.dump(recipe, default_flow_style=False, allow_unicode=True))
@@ -144,17 +195,21 @@ def run():
 			frontmatter=frontmatter,
 			content=content,
 		)
-		image = recipe.get('image')
-		if image:
-			image_url = urljoin(API, image)
-			download_image(image_url, image_folder, logger)
-		slug = slugify(recipe.get('title'))
+		slug = slugify(title)
 		filename = f'{slug}.md'
 		file_path = output_folder.joinpath(filename)
 		logger.info(f'Output path: {file_path}')
 
 		with open(file_path, 'w') as f:
 			f.write(markdown)
+		
+		add_asset(
+			asset_index, 
+			slug, 
+			title, 
+			image_urls
+		)
+	write_asset_index(asset_index)
 	logger.success('Recipes successfully imported')
